+++ CHANGELOG.md
@@ -16,6 +16,10 @@ Changelogs for this project are recorded in this file since v0.2.0.
 * Fixed centroids computations in K-shape for multivariate timeseries ([#288](https://github.com/tslearn-team/tslearn/issues/288))
 * Fixed json and pickle serialization for LearningShapelet ([#387](https://github.com/tslearn-team/tslearn/issues/387))([#403](https://github.com/tslearn-team/tslearn/issues/403)) 
 
+### Added
+
+* DBSCAN estimator for timeseries ([#598](https://github.com/tslearn-team/tslearn/issues/598))
+
 ## [v0.7.0]
 
 ### Changed